Refactor Waiver validation functions
* Move Waiver validation functions into their own module
* See #386
